+++ src/main/java/io/github/slimeistdev/semaphore/mixin/common/MixinTrain.java
@@ -18,22 +18,34 @@
 
 package io.github.slimeistdev.semaphore.mixin.common;
 
+import com.simibubi.create.content.trains.entity.Carriage;
 import com.simibubi.create.content.trains.entity.Train;
+import com.simibubi.create.content.trains.entity.TravellingPoint;
 import com.simibubi.create.content.trains.graph.DimensionPalette;
+import com.simibubi.create.content.trains.graph.TrackEdge;
 import com.simibubi.create.content.trains.graph.TrackGraph;
+import com.simibubi.create.content.trains.graph.TrackNode;
+import com.simibubi.create.content.trains.signal.SignalBoundary;
+import com.simibubi.create.content.trains.signal.TrackEdgePoint;
+import com.simibubi.create.foundation.utility.Couple;
+import com.simibubi.create.foundation.utility.Pair;
 import io.github.slimeistdev.semaphore.mixin_ducks.common.TrainDuck;
 import net.minecraft.nbt.CompoundTag;
+import org.apache.commons.lang3.mutable.MutableObject;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
 @Mixin(Train.class)
 public class MixinTrain implements TrainDuck {
+    @Shadow public List<Carriage> carriages;
     @Unique
     private boolean semaphore$navigationWatchdogDisabled = false;
 
@@ -61,6 +73,36 @@ public class MixinTrain implements TrainDuck {
             if (train != null) {
                 ((MixinTrain) (Object) train).semaphore$navigationWatchdogDisabled = true;
             }
+        }
+    }
+
+    // guard against deadlocks being caused by `/semaphore recalculate_signals`
+    @Inject(method = "lambda$collectInitiallyOccupiedSignalBlocks$20", at = @At("HEAD"), cancellable = true, remap = false)
+    private void skipLeadingSignal(
+        MutableObject<UUID> prevGroup,
+        Double distance,
+        Pair<TrackEdgePoint, Couple<TrackNode>> couple,
+        CallbackInfoReturnable<Boolean> cir
+    ) {
+        if (!(couple.getFirst() instanceof SignalBoundary signal)) return;
+
+        TravellingPoint leadingPoint = carriages.get(0).getLeadingPoint();
+        TrackNode node1 = leadingPoint.node1;
+        TrackNode node2 = leadingPoint.node2;
+        TrackEdge edge = leadingPoint.edge;
+
+        if (edge == null) return;
+
+        double position = leadingPoint.position;
+
+        if (!(Couple.create(node1, node2).equals(couple.getSecond()))) {
+            return;
+        }
+
+        double signalPosition = signal.getLocationOn(edge);
+
+        if (signalPosition + 0.0005 > position) {
+            cir.setReturnValue(false);
         }
     }
 }