+++ src/main/java/io/github/slimeistdev/semaphore/registry/SemaphoreCommandsServer.java
@@ -19,6 +19,7 @@
 package io.github.slimeistdev.semaphore.registry;
 
 import com.mojang.brigadier.CommandDispatcher;
+import io.github.slimeistdev.semaphore.content.commands.server.RecalculateSignalsCommand;
 import io.github.slimeistdev.semaphore.content.commands.server.ReloadCommandsCommand;
 import io.github.slimeistdev.semaphore.content.commands.server.TrainTeleportCommand;
 import io.github.slimeistdev.semaphore.content.commands.server.WatchdogCommand;
@@ -33,7 +34,8 @@ public class SemaphoreCommandsServer {
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher, CommandBuildContext context, Commands.CommandSelection selection) {
         var semaphoreCommand = literal("semaphore")
             .then(TrainTeleportCommand.register())
-            .then(WatchdogCommand.register());
+            .then(WatchdogCommand.register())
+            .then(RecalculateSignalsCommand.register());
 
         if (Utils.isDevEnv()) {
             semaphoreCommand = semaphoreCommand